Make GuruMapGenerator

+++ AnyGIS_filegenerator/AnyGIS_filegenerator/Model of Data/Text templates/GuruTemplates.swift
@@ -41,7 +41,7 @@ struct GuruTemplates {
     
     
     
-    func getGuruMapsItem(url: String, zoomMin: Int, zoomMax: Int, serverParts: String) -> String {
+    func getGuruMapsItem(url: String, zoomMin: Int64, zoomMax: Int64, serverParts: String) -> String {
         
         let firtstPart = """
         <layer>
